story/requirements.txt
django
python-decouple
django-debug-toolbar
django-crispy-forms
reportlab
django-ckeditor
beautifulsoup4